+++ src/components/Cart/CartItem.tsx
@@ -1,19 +1,9 @@
 import React, { useState } from 'react';
 import Link from 'next/link';
-import styled, { css } from 'styled-components';
 import api from '@utils/api';
 import { Buttons } from '@components';
 import Modal from '../Modals';
-
-interface ItemProps {
-  flex: number;
-  center?: boolean;
-  columnDirection?: boolean;
-}
-
-interface CheckBoxProps {
-  checked: boolean;
-}
+import * as Styled from './styled';
 
 export interface CartData {
   _id: number;
@@ -134,26 +124,26 @@ function CartItem({ cartData, dispatch }: CartDataProps) {
   };
 
   return (
-    <SectionItem>
-      <ContainerCheckBox>
-        <ContainerCheck checked={checked} onClick={checkToggle} />
-      </ContainerCheckBox>
-      <ContainerItem flex={55}>
-        <ImgItem src="https://itec.snu.ac.kr/msc/default.png" />
-        <ContainerText>
-          <TextCategory>{categories}</TextCategory>
+    <Styled.SectionItem>
+      <Styled.ContainerCheckBox>
+        <Styled.ContainerCheck checked={checked} onClick={checkToggle} />
+      </Styled.ContainerCheckBox>
+      <Styled.ContainerItem flex={55}>
+        <Styled.ImgItem src="https://itec.snu.ac.kr/msc/default.png" />
+        <Styled.ContainerText>
+          <Styled.TextCategory>{categories}</Styled.TextCategory>
           <p>{productName}</p>
-          <TextItemPrice>{originPrice}</TextItemPrice>
-          <TextDelivery>{deliveryCharge}</TextDelivery>
-        </ContainerText>
-      </ContainerItem>
-      <ContainerItem flex={15} center>
-        <BtnControl onClick={minusCount}>-</BtnControl>
-        <BtnCount disabled>{count}</BtnCount>
-        <BtnControl onClick={plusCount}>+</BtnControl>
-      </ContainerItem>
-      <ContainerItem flex={25} center columnDirection>
-        <TextFinalItemPrice>{price}</TextFinalItemPrice>
+          <Styled.TextItemPrice>{originPrice}</Styled.TextItemPrice>
+          <Styled.TextDelivery>{deliveryCharge}</Styled.TextDelivery>
+        </Styled.ContainerText>
+      </Styled.ContainerItem>
+      <Styled.ContainerItem flex={15} center>
+        <Styled.BtnControl onClick={minusCount}>-</Styled.BtnControl>
+        <Styled.BtnCount disabled>{count}</Styled.BtnCount>
+        <Styled.BtnControl onClick={plusCount}>+</Styled.BtnControl>
+      </Styled.ContainerItem>
+      <Styled.ContainerItem flex={25} center columnDirection>
+        <Styled.TextFinalItemPrice>{price}</Styled.TextFinalItemPrice>
         <Link
           href={{
             pathname: 'order',
@@ -173,8 +163,8 @@ function CartItem({ cartData, dispatch }: CartDataProps) {
             주문하기
           </Buttons.Custom>
         </Link>
-      </ContainerItem>
-      <AreaDelete onClick={() => setIsModal((prev) => !prev)} />
+      </Styled.ContainerItem>
+      <Styled.AreaDelete onClick={() => setIsModal((prev) => !prev)} />
       <Modal
         isModal={isModal}
         setIsModal={setIsModal}
@@ -184,128 +174,8 @@ function CartItem({ cartData, dispatch }: CartDataProps) {
       >
         정말 삭제하시겠습니까?
       </Modal>
-    </SectionItem>
+    </Styled.SectionItem>
   );
 }
-
-const SectionItem = styled.section`
-  display: flex;
-  position: relative;
-  align-items: center;
-  padding: 22px 0px;
-  border: 2px solid #e0e0e0;
-  border-radius: 10px;
-  margin-bottom: 10px;
-`;
-
-const ContainerCheckBox = styled.div`
-  flex-basis: 5%;
-`;
-
-const ContainerCheck = styled.div<CheckBoxProps>`
-  position: relative;
-  width: 20px;
-  height: 20px;
-  border: 2px solid #21bf48;
-  border-radius: 50%;
-  margin: 0 auto;
-
-  ${(props) =>
-    props.checked &&
-    css`
-      &::after {
-        content: '';
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        display: block;
-        width: 12px;
-        height: 12px;
-        background-color: #21bf48;
-        border-radius: 50%;
-      }
-    `}
-`;
-
-const ContainerItem = styled.div<ItemProps>`
-  font-size: 18px;
-  flex-basis: ${(props) => props.flex}%;
-  display: flex;
-  ${(props) => props.center && 'justify-content : center'};
-  ${(props) =>
-    props.columnDirection && 'flex-direction:column; align-items: center'};
-`;
-
-const ImgItem = styled.img`
-  display: block;
-  width: 160px;
-  height: 160px;
-  border-radius: 10px;
-`;
-
-const ContainerText = styled.div`
-  margin-left: 36px;
-`;
-
-const TextCategory = styled.em`
-  font-size: 14px;
-  color: #767676;
-  display: inline-block;
-  margin: 10px 0;
-`;
-
-const TextItemPrice = styled.p`
-  font-size: 16px;
-  font-weight: bold;
-  margin: 10px 0 40px;
-`;
-
-const TextDelivery = styled.em`
-  font-size: 14px;
-  color: #767676;
-`;
-
-const BtnControl = styled.button`
-  width: 50px;
-  height: 50px;
-  border: 1px solid #c4c4c4;
-  color: black;
-  &:first-of-type {
-    border-top-left-radius: 5px;
-    border-bottom-left-radius: 5px;
-  }
-  &:last-of-type {
-    border-top-right-radius: 5px;
-    border-bottom-right-radius: 5px;
-  }
-`;
-
-const BtnCount = styled.button`
-  width: 50px;
-  height: 50px;
-  border-top: 1px solid #c4c4c4;
-  border-bottom: 1px solid #c4c4c4;
-  color: black;
-`;
-
-const TextFinalItemPrice = styled.p`
-  margin-bottom: 18px;
-  font-size: 18px;
-  font-weight: bold;
-  color: red;
-`;
-
-const AreaDelete = styled.span`
-  display: inline-block;
-  position: absolute;
-  top: 18px;
-  right: 18px;
-  width: 22px;
-  height: 22px;
-  background: url(images/icon-delete.png) no-repeat;
-  background-size: 100%;
-  cursor: pointer;
-`;
 
 export default CartItem;
